New Threading System Implementation
JavaFX GUI had required functionalities added but is still incomplete.
Many classes were deleted in the process of creating a new threading
system located in the Moveable class. The intent is for all moveable
classes and objects to inherit their run from moveable and only require
a thread to be created for them to be run on when a destination is given
to them.