Godzilla Creation Complete
Rewrote the code for godzilla creation; it now works! Also, added
functionality to mapMove so that, if given a null location, it will
remove the inputed Move from the map and redraw any Move that was
occupying the same ship. Hopefully, this funcitonality can be used to
safely remove single objects from the map without redraw bugs or ghost
images.